SCALE-248: wip
