+++ Gemfile
@@ -1,4 +1,5 @@
-source "https://rubygems.org"
+source 'https://rubygems.org'
+source 'https://gem.fury.io/chime/'
 
 git_source(:github) {|repo_name| "https://github.com/#{repo_name}" }
