+++ lib/simplekiq/metadata_client.rb
@@ -1,88 +1,95 @@
+require 'chime_atlas'
 require 'sidekiq'
+require 'socket'
+require 'time'
+require 'pry'
 
 module Simplekiq
   class MetadataClient
     def config
       Sidekiq.configure_server do |config|
         config.client_middleware do |chain|
-          chain.add(MetadataClient)
+          chain.add(Simplekiq::MetadataClient)
         end
       end
     end
 
-    def call(worker, job, queue, *)
+    def call(_worker, job, _queue, *)
+      binding.pry
       begin
-        record_preexecute(job)
+        record(job)
         yield
-      rescue e
-        record_error(job, e)
-      ensure
-        record_postexecute(job)
-        dispatch_metadata_callback(worker, job, queue)
       end
     end
 
-    def record_preexecute(job)
-      job[:_metadata] |= {}
-      job[:_metadata].merge(client_preexecute_metadata(job))
+    def record(job)
+      job['_metadata'] = server_metadata
     end
 
-    def record_postexecute(job)
-      job[:_metadata] |= {}
-      job[:_metadata].merge(client_postexecute_metadata(job))
-    end
-
-    def record_error(job, e)
-      job[:_metadata][:error] = {
-        message: e.message,
-        trace: e.backtrace.inspect
-      }
-    end
-
-    def client_preexecute_metadata(job)
-      now = processed_at
+    def server_metadata
       {
-        first_processed_at: first_processed_at(job, now),
-        processed_at: processed_at,
-        processed_by: processed_by,
-        processed_by_host: processed_by_host
+        enqueued_at: enqueued_at,
+        enqueued_from: enqueued_from,
+        enqueued_from_host: enqueued_from_host,
+        enqueued_from_src_location: enqueued_from_src_location,
+        request_id: request_id
       }
     end
 
-    def client_postexecute_metadata(job)
-      {
-        retries: retries(job)
-      }
+    def request_id
+      Chime::Atlas.request_id
     end
 
-    def first_processed_at(job, processed_at)
-      first_processed_at = job[:_metadata][:first_processed_at]
-      if first_processed_at.nil? || first_processed_at.empty?
-        first_processed_at = processed_at
-      end
-      first_processed_at
-    end
-
-    def processed_at
+    def enqueued_at
       Time.now.utc.iso8601
     end
 
-    def processed_by
-      @processed_service ||= Chime::Dog.config['app']
+    def enqueued_from
+      @enqueued_service ||= Chime::Dog.config['app']
     end
 
-    def processed_by_host
-      @processed_host ||= Socket.gethostname
+    def enqueued_from_host
+      @enqueued_host ||= Socket.gethostname
     end
 
-    def retries(job)
-      return 0 if job[:_metadata][:retries].nil? or job[:_metadata][:retries].empty?
+    def enqueued_from_src_location
+      last_stack_is_simplekiq = false
+      caller.inspect.reverse_each do |callstack_line|
+        entry = split_stackline(callstack_line)
+        if is_simplekiq_callstack?(entry)
+          last_stack_is_simplekiq = true
+        else
+          if last_stack_is_simplekiq
+            return callstack_entry_hash(entry)
+          end
+          last_stack_is_simplekiq = false
+        end
+      end
 
-      job[:_metadata][:retries] + 1
+      callstack_entry_hash([nil, nil, nil])
     end
 
-    def dispatch_metadata_callback(worker, job, queue)
-      nil
+    private
+
+    def is_simplekiq_callstack?(callstack_entry)
+      callstack_entry.first.include?('simplekiq') || callstack_entry.first.include?('sidekiq')
+    end
+
+    def callstack_entry_hash(callstack_entry)
+      {
+        path: callstack_entry[0],
+        line: callstack_entry[1],
+        method: callstack_entry[2]
+      }
+    end
+
+    def split_stackline(callstack_line)
+      split = callstack_line.rpartition(':in ')
+      method = split.last.tr('`', '')
+      split_2 = split.first.rpartition(':')
+      path = split_2.first
+      line = split_2.last
+      [path, line, method].freeze
     end
   end
 end
